+++ core/src/test/java/org/apache/struts/config/TestModuleConfig.java
@@ -27,7 +27,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 
 import java.io.InputStream;
 
-import org.apache.commons.digester3.Digester;
+import org.apache.commons.digester.Digester;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;