+++ tiles/src/main/java/org/apache/struts/tiles/xmlDefinition/XmlParser.java
@@ -27,7 +27,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
 
-import org.apache.commons.digester3.Digester;
+import org.apache.commons.digester.Digester;
 import org.xml.sax.SAXException;
 
 /**